Generate & Evaluate Embedding

get_eval_embed acts as embedding generation & evaluation from co-occurrence data file. It returns a list of summary including meta-data, evaluation and embedding itself.

get_eval_embed(
  CO_file,
  ARP_file,
  dims,
  CO_dict_file = NULL,
  out_dir = NULL,
  HAM_file = NULL,
  data_type = 1
)

Arguments

CO_file

Co-ccurrence data file with format .csv, .parquet or .Rdata. The data should be a table with 3 columns V1, V2, V3:

V1: Shows the row id (code id) or row code pair.

V2: Shows the col id (code id) or row code pair..

V3: Shows the counts for certain pair.

Note: If V1 & V2 are code ids other than code pairs, a mapping dict offering information from id to code pair is needed. Please see CO_dict_file for details.

ARP_file

All relation pairs data file with format .csv, .parquet or .Rdata.

dims

A vector of numeric values for dimension.

CO_dict_file

A data file with two columns, where:

Column 1: Shows the name of code pair.

Column 2: Shows the corresponding id number.

File format can be .csv, .parquet or .Rdata.

out_dir

Output folder, if NULL then by default set to your_working_directory/output.

HAM_file

Multi-axial hierarchy data file with format .csv, .parquet or .Rdata. By default it's set to NULL and read the default file within the package. If a file name is offered, it'll read it instead and replace the default file

data_type

If data Does not contain CUI codes, set as 1. Otherwise, set as 2.

Value

A list of infomation of meta-data, embedding & evaluation result. It will be saved in out_dir as .Rdata file.