Generate Plot Report

get_report acts as plot report generation from summary returned by get_eval_embed. It returns NULL but generates plots in out_dir, classified by similarity & relation. Report format can be chosen in parameter knit_format.

Usage

get_report(
  summary,
  plot_val = "auc",
  knit_format = "html",
  labels = c("PheCode-PheCode(sim)", "RXNORM-RXNORM(sim)", "LAB-LAB(sim)",
    "PheCode-PheCode(rela)", "PheCode-RXNORM(rela)", "PheCode-LAB(rela)"),
  split_patterns = list(Similarity = "(sim)", Relation = "(rela)")
)

Arguments

summary

A list of summary infomation generated by get_eval_embed.

plot_val

Metric for evaluation (Columns in summary evaluation dataframes). Options could be: #"pairs", "auc", "cut/0.01", "cut/0.05", "cut/0.1", "TPR/0.01", "TPR/0.05" or "TPR/0.1".

knit_format

Format of plot report, "pdf" or "html".

labels

Labels to be used in plots, options are the row names of summary evaluation dataframes.

split_patterns

Pattern to split lablels, normally not changed.